Reworked how latest change detection works when syncing the query store. Instead of filtering for events which have UtcDateLastModified > current update date, we now filter for the event Id instead.
This is needed because in some rare cases (= typically only during testing) when events are fired very close to each other, sometimes the timestamp resolution is not fine enough and some DB records which were actually caused by different events will have the same timestamp. When that occurs, this is what happens:
- Add Tag "Dev" and "QA" to ticket 1
- These get stored in DB. Since they are stored in the same transaction, they will have the same timestamp, but that's fine, because when the save occurs, we send a notification and the query store synchronizer will pick both of them up fine.
- The query store sync updates the last event date.
- Remove Tag "Dev" from ticket 1
- If this runs fast enough (yes, it can happen) then this will have the exact same timestamp up to the smallest fraction which the DB can store.
- So when the query store sync runs, it finds that the data in the store is up-to-date and will skip the sync.

This should not happen in typical end-user scenarios but can happen when, for whatever reason, two or more distinct commands are processed close enough to each other. A fabricated example for this can be fond in the TicketCommandsController file as of this commit, where (currently) there are test commands being raised.